requirements.txt
python-dotenv==1.1.0
bcrypt==4.3.0